update tests, remove cross-references to infocontent test.
